Inclusão da UserControllerTest

+++ projeto-biblioteca/Services/interface/IReservationService.cs
@@ -2,7 +2,7 @@ using pBiblioteca.Models;
 public interface IReservationService
 {
     string CreateReservation(CreateReservationRequest request);
-    string CancelReservation(long reservationId);
+    string CancelReservation(int reservationId);
     int GetUserPosition(string isbn, string cpf);
     List<ReservationResponseDTO> GetQueue(string isbn);
 }
